pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kanji-clustering-api"
version = "0.1.0"
description = "A FastAPI API for clustering kanji characters"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=3.6.2",
    "asgiref==3.5.2",
    "astroid==2.11.5",
    "autopep8==1.6.0",
    "certifi==2024.7.4",
    "cffi==1.15.0",
    "chardet==3.0.4",
    "click==8.1.3",
    "deprecated==1.2.13",
    "deta==1.1.0",
    "dill==0.3.5.1",
    "fastapi==0.116.1",
    "flake8==4.0.1",
    "flask==3.0.2",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "idna==3.7",
    "imageio==2.19.3",
    "isort==5.10.1",
    "itsdangerous==2.1.2",
    "jedi==0.17.2",
    "jinja2==3.1.5",
    "joblib==1.2.0",
    "lazy-object-proxy==1.7.1",
    "mangum==0.15.0",
    "markupsafe==2.1.1",
    "mccabe==0.6.1",
    "mypy-extensions==0.4.3",
    "mypy>=1.0.0",
    "networkx==2.8.3",
    "numpy>1.22.3",
    "oauthlib==3.1.0",
    "packaging==21.3",
    "pandas==1.4.2",
    "parso==0.7.1",
    "pillow==10.3.0",
    "platformdirs==2.5.2",
    "pluggy==1.0.0",
    "progressbar2==4.0.0",
    "pycodestyle==2.8.0",
    "pycparser==2.21",
    "pydantic==1.10.13",
    "pydocstyle==6.1.1",
    "pyflakes==2.4.0",
    "pyjwt==2.4.0",
    "pylint==2.13.9",
    "pynacl==1.5.0",
    "pyparsing==3.0.9",
    "python-dateutil==2.8.2",
    "python-jsonrpc-server==0.4.0",
    "python-language-server==0.36.2",
    "python-lsp-jsonrpc==1.0.0",
    "python-lsp-server==1.4.1",
    "python-utils==3.2.3",
    "pytz==2022.1",
    "pywavelets==1.7.0",
    "requests-oauthlib==1.3.0",
    "requests==2.32.4",
    "rope==1.1.1",
    "scikit-image>0.19.2",
    "scikit-learn>1.1.1",
    "scipy==1.14.1",
    "six==1.16.0",
    "sniffio==1.2.0",
    "snowballstemmer==2.2.0",
    "starlette==0.47.1",
    "threadpoolctl==3.1.0",
    "tifffile==2022.5.4",
    "toml==0.10.2",
    "tomli==2.0.1",
    "typing-extensions>=4.10.0",
    "ujson==5.9.0",
    "urllib3==2.5.0",
    "uvicorn==0.17.6",
    "werkzeug==3.0.6",
    "wrapt==1.14.1",
    "xlrd==2.0.1",
    "yapf==0.32.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.4",
    "types-requests",
    "pandas-stubs",
]

[dependency-groups]
lint = [
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]



[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",  # blank line before class docstring (conflicts with D211)
    "D213",  # multi-line summary should start at second line (conflicts with D212)
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "progressbar",
    "sklearn.cluster",
    "uvicorn",
]
ignore_missing_imports = true
